requirements.txt
certifi==2019.6.16
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.5
django-cors-headers==3.0.2
django-heroku==0.3.1
djangorestframework==3.9.1
gunicorn==19.9.0
idna==2.8
mysqlclient==1.4.2.post1
psycopg2==2.8.3
pytz==2019.2
requests==2.21.0
urllib3==1.24.3
whitenoise==4.1.3
